exploded prj_commands into the project_commands module.
