+++ python/tk_desktop/project_commands/default_grouping_header.py
@@ -1,0 +1,47 @@
+# Copyright (c) 2020 Shotgun Software Inc.
+#
+# CONFIDENTIAL AND PROPRIETARY
+#
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
+# Source Code License included in this distribution package. See LICENSE.
+# By accessing, using, copying or modifying this work you indicate your
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
+# not expressly granted therein are reserved by Shotgun Software Inc.
+
+from sgtk.platform.qt import QtCore, QtGui
+
+
+class DefaultGroupingHeader(QtGui.QPushButton):
+    """ Default widget for a group header """
+
+    def __init__(self, parent=None):
+        QtGui.QPushButton.__init__(self, parent)
+
+        # cache the icons for collapsed/expanded
+        self.down_arrow = QtGui.QIcon(":tk-desktop/down_arrow.png")
+        self.right_arrow = QtGui.QIcon(":tk-desktop/right_arrow.png")
+
+        # adjust the button look
+        self.setFlat(True)
+        self.setCheckable(True)
+        self.setChecked(True)
+        self.setFocusPolicy(QtCore.Qt.NoFocus)
+        self.setStyleSheet(
+            """
+            text-align: left;
+            font-size: 14px;
+            background-color: transparent;
+            border: none;
+        """
+        )
+
+    def is_expanded(self):
+        return self.isChecked()
+
+    def set_expanded(self, is_expanded):
+        """ Set the widget to be expanded or collapsed """
+        self.setChecked(is_expanded)
+        if is_expanded:
+            self.setIcon(self.down_arrow)
+        else:
+            self.setIcon(self.right_arrow)
