+++ tests/run_project_commands.py
@@ -34,7 +34,7 @@ sgtk.platform.qt.QtCore = importer.QtCore
 QtGui = sgtk.platform.qt.QtGui
 QtCore = sgtk.platform.qt.QtCore
 
-from tk_desktop.prj_commands import CommandsView
+from tk_desktop.prj_commands import ProjectCommands
 from tk_desktop.ui.desktop_window import Ui_DesktopWindow
 
 
@@ -106,7 +106,7 @@ main.ui.setupUi(main)
 # Change the current page so the project commands page is visible.
 main.ui.apps_tab.setCurrentIndex(1)
 
-view = CommandsView(main.ui.project_commands_area, ProjectCommandSettings())
+view = ProjectCommands(main.ui.project_commands_area, ProjectCommandSettings())
 main.ui.project_commands_area.setWidget(view)
 
 view.set_project(
