+++ tests/test_project_commands.py
@@ -22,7 +22,8 @@ sgtk.platform.qt.QtGui = importer.QtGui
 sgtk.platform.qt.QtCore = importer.QtCore
 
 
-from prj_commands import CommandsView, RecentList
+from project_commands import ProjectCommands
+from project_commands.recent_list import RecentList
 
 PROJECT = {"type": "Project", "id": 3}
 PROJECT_KEY = "project_recent_apps.3"
@@ -35,7 +36,7 @@ def qapplication():
 
 @pytest.fixture
 def simple_test_view():
-    view = CommandsView(sgtk.platform.qt.QtGui.QScrollArea(), Settings())
+    view = ProjectCommands(sgtk.platform.qt.QtGui.QScrollArea(), Settings())
     view.set_project(PROJECT, ["Creative Tools", "Editorial"])
     return view
 
@@ -71,7 +72,7 @@ def test_sections_sorted(show_recents, commands):
     """
     groups = ["Studio", "Creative Tools", "Editorial", "Automotive Tools"]
     # Create a view with some recents.
-    view = CommandsView(
+    view = ProjectCommands(
         sgtk.platform.qt.QtGui.QScrollArea(),
         Settings(
             {PROJECT_KEY: {"command 0": {"timestamp": datetime.datetime.utcnow()}}}
@@ -110,7 +111,7 @@ def test_clear_deletes_all_but_stretcher():
     """
     Ensure clearing removes all widget except for the stretcher
     """
-    view = CommandsView(
+    view = ProjectCommands(
         sgtk.platform.qt.QtGui.QScrollArea(),
         Settings(
             {PROJECT_KEY: {"maya_2020": {"timestamp": datetime.datetime.utcnow()}}}
@@ -244,7 +245,7 @@ def test_recent_sorted_properly(recents, monkeypatch):
         }
     )
     monkeypatch.setattr(RecentList, "MAX_RECENTS", 3)
-    view = CommandsView(sgtk.platform.qt.QtGui.QScrollArea(), settings)
+    view = ProjectCommands(sgtk.platform.qt.QtGui.QScrollArea(), settings)
     view.set_project(PROJECT, ["Creative Tools"], show_recents=True)
     _register_commands(view, commands)
     assert _get_recents_title(view) == [
@@ -264,7 +265,7 @@ def _get(iterator, position):
 
 def test_recent_time_update_when_clicking():
     settings = Settings()
-    view = CommandsView(sgtk.platform.qt.QtGui.QScrollArea(), settings)
+    view = ProjectCommands(sgtk.platform.qt.QtGui.QScrollArea(), settings)
     view.set_project(PROJECT, ["Creative Tools"], show_recents=True)
     _register_commands(view, ["Maya 2017", "Maya 2018*", "Maya 2019"])
